refactor: split layout.tsx
