--- vrc-get-gui/app/(main)/fonts.ts
@@ -1,5 +1,0 @@
-import {Noto_Sans_Mono} from "next/font/google";
-
-export const notoSansMono = Noto_Sans_Mono({
-	subsets: ["latin"]
-});
