+++ vrc-get-gui/app/(main)/layout.tsx
@@ -1,29 +1,16 @@
-import type {Metadata} from "next";
-import "./globals.css";
-import 'react-toastify/dist/ReactToastify.css';
 import {SideBar} from "@/components/SideBar";
-import {Providers} from "@/components/providers";
 
-export const metadata: Metadata = {
-	title: "Create Next App",
-	description: "Generated by create next app",
-};
-
-export default function RootLayout({
+export default function MainLayout({
 																		 children,
 																	 }: Readonly<{
 	children: React.ReactNode;
 }>) {
 	return (
-		<html>
-		<body className={`font-sans w-screen h-screen flex flex-row overflow-hidden whitespace-nowrap`}>
-		<Providers>
+		<>
 			<SideBar className={"flex-grow-0"}/>
 			<div className={"h-screen flex-grow overflow-hidden flex p-4"}>
 				{children}
 			</div>
-		</Providers>
-		</body>
-		</html>
+		</>
 );
 }
