+++ vrc-get-gui/app/layout.tsx
@@ -1,0 +1,25 @@
+import type {Metadata} from "next";
+import "./globals.css";
+import 'react-toastify/ReactToastify.css';
+import {Providers} from "@/components/providers";
+
+export const metadata: Metadata = {
+	title: "Create Next App",
+	description: "Generated by create next app",
+};
+
+export default function RootLayout({
+																		 children,
+																	 }: Readonly<{
+	children: React.ReactNode;
+}>) {
+	return (
+		<html>
+		<body className={`font-sans w-screen h-screen flex flex-row overflow-hidden whitespace-nowrap`}>
+		<Providers>
+			{children}
+		</Providers>
+		</body>
+		</html>
+);
+}
